requirements.txt
selenium
python-dotenv
random-username
fake-useragent